+++ src/main/java/com/exchanger/currency/integration/currency/CurrencyResponse.java
@@ -1,6 +1,7 @@
 package com.exchanger.currency.integration.currency;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
-public record CurrencyResponse(String currency, String code, double mid, LocalDate date) {
+public record CurrencyResponse(String currency, String code, BigDecimal mid, LocalDate date) {
 }
