+++ src/main/java/com/exchanger/currency/services/excel/CurrencyReportCurrencies.java
@@ -2,7 +2,9 @@ package com.exchanger.currency.services.excel;
 
 import com.exchanger.currency.domain.currency.Currency;
 
-public record CurrencyReportCurrencies(String name, String code, Double mid) {
+import java.math.BigDecimal;
+
+public record CurrencyReportCurrencies(String name, String code, BigDecimal mid) {
     public static CurrencyReportCurrencies from(Currency currency){
         return new CurrencyReportCurrencies(currency.getCurrency(),
                 currency.getCode(),
